backend/requirements.txt
absl-py==1.4.0
aioboto3==11.3.0
aiobotocore==2.6.0
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
ansi2html==1.8.0
anyio==4.6.2.post1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
aws-cdk-lib==2.89.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.166
aws-cdk.aws-lambda-python-alpha==2.17.0a0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.11.1
blinker==1.6.2
boto3==1.28.17
botocore==1.31.17
Brotli==1.0.9
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
chart-studio==1.1.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
constructs==10.2.69
contourpy==1.1.0
cycler==0.11.0
dash==2.8.1
dash-auth==1.4.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7.post1
decorator==5.1.1
df2gspread==1.0.4
distlib==0.3.6
dnspython==1.16.0
ecdsa==0.19.0
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.115.2
filelock==3.9.0
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Compress==1.13
Flask-SeaSurf==1.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
gast==0.4.0
google-api-python-client==1.6.7
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.57.0
gspread==5.6.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.0.0
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsii==1.85.0
jupyter_client==8.3.1
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.7
nltk==3.8.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
passlib==1.7.4
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.17.0
pluggy==1.2.0
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
publication==0.0.3
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.16.1
pymongo==3.11.0
pyparsing==3.0.9
pypng==0.20220715.0
pytest==7.4.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.12
pytz==2023.3
pywin32==306
pyzmq==25.1.1
qrcode==7.4.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.2.post1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.22
stack-data==0.6.2
starlette==0.40.0
statsmodels==0.14.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
textblob==0.17.1
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2023.3
ua-parser==0.16.1
uritemplate==3.0.1
urllib3==1.26.16
uvicorn==0.32.0
vaderSentiment==3.3.2
virtualenv==20.20.0
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.13.0
